Reduce MCP server integration tests
