+++ Rakefile
@@ -22,7 +22,12 @@ Rake::TestTask.new(ruby_test: :compile, &test_config)
 
 begin
   require "ruby_memcheck"
-  namespace(:ruby_test) { RubyMemcheck::TestTask.new(valgrind: :compile, &test_config) }
+  namespace(:ruby_test) do
+    RubyMemcheck::TestTask.new(valgrind: :compile) do |task|
+      test_config.call(task)
+      task.test_files = FileList["test/**/*_test.rb"].exclude("test/integration/**/*_test.rb")
+    end
+  end
 rescue LoadError
   # ruby_memcheck is not available on Windows
 end
