+++ test/mcp_server_test.rb
@@ -5,8 +5,6 @@ require "helpers/context"
 require "json"
 require "mocha/minitest"
 require "rubydex/mcp_server"
-require "open3"
-require "rbconfig"
 require "stringio"
 require "timeout"
 require "uri"
@@ -301,289 +299,5 @@ class MCPServerTest < Minitest::Test
     result = response.fetch(:result)
     assert_equal(true, result.fetch(:isError))
     assert_equal("Missing required arguments: query", result.fetch(:content)[0].fetch(:text))
-  end
-end
-
-class MCPServerIntegrationTest < Minitest::Test
-  include Test::Helpers::WithContext
-
-  MAX_INDEXING_RETRIES = 200
-
-  def test_executable_prints_help
-    stdout, _stderr, status = run_executable("--help")
-
-    assert_predicate(status, :success?)
-    assert_includes(stdout, "mcp [PATH]")
-    assert_includes(stdout, "Run the MCP server for AI assistants")
-  end
-
-  def test_executable_prints_version
-    stdout, _stderr, status = run_executable("--version")
-
-    assert_predicate(status, :success?)
-    assert_equal("v#{Rubydex::VERSION}\n", stdout)
-  end
-
-  def test_executable_rejects_extra_arguments
-    stdout, stderr, status = run_executable("mcp", "foo", "bar")
-
-    assert_equal(1, status.exitstatus)
-    assert_empty(stdout)
-    assert_includes(stderr, "unexpected argument: bar")
-    assert_includes(stderr, "mcp [PATH]")
-  end
-
-  def test_executable_rejects_unknown_options
-    stdout, stderr, status = run_executable("mcp", "--unknown")
-
-    assert_equal(1, status.exitstatus)
-    assert_empty(stdout)
-    assert_includes(stderr, "invalid option: --unknown")
-    assert_includes(stderr, "mcp [PATH]")
-  end
-
-  def test_mcp_server_can_be_required_directly
-    stdout, stderr, status = Open3.capture3(
-      RbConfig.ruby,
-      "-rbundler/setup",
-      "-Ilib",
-      "-e",
-      <<~RUBY,
-        require "rubydex/mcp_server"
-
-        puts Rubydex::VERSION
-        puts Rubydex::Graph.name
-
-        response = Rubydex::MCPServer::Server.new(root_path: Dir.pwd).handle(
-          jsonrpc: "2.0",
-          id: 1,
-          method: "initialize",
-        )
-        puts response.fetch(:result).fetch(:serverInfo).fetch(:version)
-      RUBY
-    )
-
-    assert_predicate(status, :success?, stderr)
-    assert_equal("#{Rubydex::VERSION}\nRubydex::Graph\n#{Rubydex::VERSION}\n", stdout)
-  end
-
-  def test_mcp_server_e2e
-    skip("This external process e2e times out when ruby_memcheck traces child processes under Valgrind") if ENV["RUBY_MEMCHECK_RUNNING"]
-
-    with_context do |context|
-      context.write!("app.rb", <<~RUBY)
-        class Animal
-          def speak
-            "..."
-          end
-        end
-
-        class Dog < Animal
-          def speak
-            "Woof!"
-          end
-        end
-
-        module Greetable
-          def greet
-            "Hello"
-          end
-        end
-
-        module UniqueMarker
-        end
-
-        class Kennel
-          def build
-            Animal.new
-          end
-        end
-
-      RUBY
-
-      stderr_output = +""
-      Open3.popen3(RbConfig.ruby, "-rbundler/setup", executable_path, "mcp", context.absolute_path) do |stdin, stdout, stderr, wait_thr|
-        stderr_reader = Thread.new do
-          stderr_output << stderr.read
-        rescue IOError
-          # This thread only drains stderr. If teardown closes the pipe while read is blocked, there is
-          # no more stderr to capture and the subprocess exit status is still checked through wait_thr.
-        end
-
-        initialize_session(stdin, stdout)
-        assert_tools_are_registered(stdin, stdout)
-
-        request_id = 3
-        stats, request_id = wait_for_indexing_to_complete(stdin, stdout, request_id)
-        assert_operator(stats.fetch("files"), :>=, 2)
-        assert_operator(stats.fetch("declarations"), :>, 0)
-
-        request_id += 1
-        search_response = call_tool(stdin, stdout, request_id, "search_declarations", { query: "Dog", match_mode: "exact", kind: "Class" })
-        assert_has_name(search_response.fetch("results"), "Dog", "search results")
-        assert_operator(search_response.fetch("total"), :>, 0)
-
-        request_id += 1
-        negative_offset_response = call_tool(stdin, stdout, request_id, "search_declarations", { query: "UniqueMarker", match_mode: "exact", offset: -1, limit: 1 })
-        assert_has_name(negative_offset_response.fetch("results"), "UniqueMarker", "negative offset search results")
-
-        request_id += 1
-        declaration = call_tool(stdin, stdout, request_id, "get_declaration", { name: "Dog" })
-        assert_equal("Dog", declaration.fetch("name"))
-        assert_equal("Class", declaration.fetch("kind"))
-        refute_empty(declaration.fetch("definitions"))
-        assert_has_name(declaration.fetch("ancestors"), "Animal", "Dog ancestors")
-
-        request_id += 1
-        descendants = call_tool(stdin, stdout, request_id, "get_descendants", { name: "Animal" })
-        assert_has_name(descendants.fetch("descendants"), "Dog", "Animal descendants")
-        assert_operator(descendants.fetch("total"), :>, 0)
-
-        request_id += 1
-        references = call_tool(stdin, stdout, request_id, "find_constant_references", { name: "Animal" })
-        refute_empty(references.fetch("references"))
-        assert(references.fetch("references").all? { |entry| entry.key?("path") })
-        assert_operator(references.fetch("total"), :>, 0)
-
-        request_id += 1
-        method_references = call_tool(stdin, stdout, request_id, "find_constant_references", { name: "Dog#speak()" })
-        assert_equal("Dog#speak()", method_references.fetch("name"))
-        assert_empty(method_references.fetch("references"))
-        assert_equal(0, method_references.fetch("total"))
-
-        request_id += 1
-        file_declarations = call_tool(stdin, stdout, request_id, "get_file_declarations", { file_path: "app.rb" })
-        assert(file_declarations.fetch("file").end_with?("app.rb"))
-        declaration_entries = file_declarations.fetch("declarations")
-        assert_has_name(declaration_entries, "Animal", "file declarations")
-        assert_has_name(declaration_entries, "Dog", "file declarations")
-        assert_has_name(declaration_entries, "Greetable", "file declarations")
-
-        stdin.close
-        Timeout.timeout(30) { wait_thr.value }
-        stderr_reader.join
-      rescue Timeout::Error
-        Process.kill("TERM", wait_thr.pid)
-        flunk("rdx mcp did not exit after stdin closed. stderr:\n#{stderr_output}")
-      end
-    end
-  end
-
-  private
-
-  def run_executable(*arguments)
-    Open3.capture3(
-      RbConfig.ruby,
-      "-rbundler/setup",
-      executable_path,
-      *arguments,
-    )
-  end
-
-  def executable_path
-    File.expand_path("../exe/rdx", __dir__)
-  end
-
-  def send_message(stdin, message)
-    stdin.puts(JSON.generate(message))
-    stdin.flush
-  end
-
-  def send_request(stdin, id, method, params)
-    send_message(
-      stdin,
-      {
-        jsonrpc: "2.0",
-        id: id,
-        method: method,
-        params: params,
-      },
-    )
-  end
-
-  def read_response(stdout)
-    Timeout.timeout(5) do
-      line = stdout.gets
-      flunk("Expected JSON-RPC response, got EOF") unless line
-
-      JSON.parse(line)
-    end
-  end
-
-  def read_response_for_id(stdout, expected_id)
-    response = read_response(stdout)
-    assert_equal(expected_id, response.fetch("id"))
-    response
-  end
-
-  def initialize_session(stdin, stdout)
-    send_request(
-      stdin,
-      1,
-      "initialize",
-      {
-        protocolVersion: "2025-03-26",
-        capabilities: {},
-        clientInfo: { name: "test-client", version: "0.1.0" },
-      },
-    )
-
-    response = read_response_for_id(stdout, 1)
-    assert_kind_of(Hash, response.fetch("result").fetch("capabilities").fetch("tools"))
-
-    send_message(
-      stdin,
-      {
-        jsonrpc: "2.0",
-        method: "notifications/initialized",
-      },
-    )
-  end
-
-  def assert_tools_are_registered(stdin, stdout)
-    send_request(stdin, 2, "tools/list", {})
-    response = read_response_for_id(stdout, 2)
-    tool_names = response.fetch("result").fetch("tools").map { |tool| tool.fetch("name") }
-
-    assert_includes(tool_names, "search_declarations")
-    assert_includes(tool_names, "get_declaration")
-    assert_includes(tool_names, "get_descendants")
-    assert_includes(tool_names, "find_constant_references")
-    assert_includes(tool_names, "get_file_declarations")
-    assert_includes(tool_names, "codebase_stats")
-    assert_equal(6, tool_names.length)
-  end
-
-  def call_tool(stdin, stdout, request_id, tool_name, arguments)
-    send_request(
-      stdin,
-      request_id,
-      "tools/call",
-      {
-        name: tool_name,
-        arguments: arguments,
-      },
-    )
-
-    response = read_response_for_id(stdout, request_id)
-    JSON.parse(response.fetch("result").fetch("content")[0].fetch("text"))
-  end
-
-  def wait_for_indexing_to_complete(stdin, stdout, request_id)
-    MAX_INDEXING_RETRIES.times do
-      parsed = call_tool(stdin, stdout, request_id, "codebase_stats", {})
-      return [parsed, request_id] unless parsed.key?("error")
-
-      assert_equal("indexing", parsed.fetch("error"))
-      request_id += 1
-      sleep(0.05)
-    end
-
-    flunk("Timed out waiting for indexing to complete")
-  end
-
-  def assert_has_name(entries, expected_name, context)
-    names = entries.filter_map { |entry| entry["name"] }
-    assert_includes(names, expected_name, "Expected #{context} to include #{expected_name}, got: #{names.inspect}")
   end
 end
